requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-filter==23.5
sqlparse==0.4.4
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
python-dotenv==1.0.1
